components/inference-with-onnx-runtime/requirements.txt
numpy==1.16.0
onnx
onnxruntime
azure-storage-blob